split HTML formatter tests into smaller files
