+++ src/Microsoft.DotNet.Interactive.Formatting.Tests/PocketViewTests.cs
@@ -1,7 +1,6 @@
 // Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Collections.Generic;
 using FluentAssertions;
 using System.Linq;
