+++ src/Microsoft.DotNet.Interactive.Formatting/MemberAccessor{T}.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
